Add_User_To_Company

+++ MyLifeJob.Business/Services/Implements/CompanyService.cs
@@ -1,13 +1,18 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using MyLifeJob.Business.Constants;
 using MyLifeJob.Business.Dtos.CompanyDtos;
 using MyLifeJob.Business.Exceptions.Commons;
+using MyLifeJob.Business.Exceptions.Company;
 using MyLifeJob.Business.Exceptions.FileService;
 using MyLifeJob.Business.Extensions;
 using MyLifeJob.Business.ExternalServices.Interfaces;
 using MyLifeJob.Business.Services.Interfaces;
 using MyLifeJob.Core.Entity;
 using MyLifeJob.DAL.Repositories.Interfaces;
+using System.Security.Claims;
 
 namespace MyLifeJob.Business.Services.Implements;
 
@@ -17,18 +22,31 @@ public class CompanyService : ICompanyService
     readonly IMapper _map;
     readonly IFileService _file;
     readonly IIndustiryRepository _industry;
+    private string? _userId;
+    readonly IHttpContextAccessor _accessor;
+    readonly UserManager<AppUser> _user;
 
-    public CompanyService(ICompanyRepository repo, IMapper map, IFileService file, IIndustiryRepository industry)
+    public CompanyService(ICompanyRepository repo, IMapper map, IFileService file, IIndustiryRepository industry,
+        IHttpContextAccessor accessor, UserManager<AppUser> user)
     {
         _repo = repo;
         _map = map;
         _file = file;
         _industry = industry;
+        _accessor = accessor;
+        _userId = _accessor.HttpContext?.User?.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        _user = user;
     }
 
     public async Task CreateAsync(CompanyCreateDto dto)
     {
+        if (string.IsNullOrEmpty(_userId)) throw new ArgumentNullException();
+        var user = await _user.Users.Include(u => u.Company).SingleOrDefaultAsync(u => u.Id == _userId);
+        if (user == null) throw new NotFoundException<AppUser>();
+
         if (await _repo.IsExistAsync(c => c.Name == dto.Name)) throw new IsExistException<Company>();
+
+        if (user.Company != null) throw new UserCompanyNotEmptyException();
 
         var map = _map.Map<Company>(dto);
 
@@ -51,15 +69,22 @@ public class CompanyService : ICompanyService
             });
         }
         map.CompanyIndustries = ci;
+        map.AppUserId = _userId;
         await _repo.CreateAsync(map);
         await _repo.SaveAsync();
     }
 
     public async Task DeleteAsync(int id)
     {
+        if (string.IsNullOrEmpty(_userId)) throw new ArgumentNullException();
+        var user = await _user.Users.Include(a => a.Company).SingleOrDefaultAsync(a => a.Id == _userId);
+        if (user == null) throw new NotFoundException<AppUser>();
+
         if (id < 0) throw new IdIsNegativeException<Company>();
-        var entity = await _repo.FindByIdAsync(id);
-        if(entity == null) throw new NotFoundException<Company>();
+        var entity = await _repo.FindByIdAsync(id, "AppUser");
+        if (entity == null) throw new NotFoundException<Company>();
+
+        if (entity.AppUserId != _userId) throw new UserNotSameException();
 
         if (entity.Logo != null) _file.Delete(entity.Logo);
         await _repo.DeleteAsync(id);
@@ -70,11 +95,11 @@ public class CompanyService : ICompanyService
     {
         if (takeAll == true)
         {
-            return _map.Map<ICollection<CompanyListItemDto>>(_repo.GetAllAsync("CompanyIndustries", "CompanyIndustries.Industry"));
+            return _map.Map<ICollection<CompanyListItemDto>>(_repo.GetAllAsync("CompanyIndustries", "CompanyIndustries.Industry", "AppUser"));
         }
         else
         {
-            return _map.Map<ICollection<CompanyListItemDto>>(_repo.FindAllAsync(c => c.IsDeleted == false, "CompanyIndustries", "CompanyIndustries.Industry"));
+            return _map.Map<ICollection<CompanyListItemDto>>(_repo.FindAllAsync(c => c.IsDeleted == false, "CompanyIndustries", "CompanyIndustries.Industry", "AppUser"));
         }
     }
 
@@ -84,13 +109,13 @@ public class CompanyService : ICompanyService
 
         if (takeAll == true)
         {
-            var entity = await _repo.FindByIdAsync(id, "CompanyIndustries", "CompanyIndustries.Industry");
+            var entity = await _repo.FindByIdAsync(id, "CompanyIndustries", "CompanyIndustries.Industry", "AppUser");
             if (entity == null) throw new NotFoundException<Company>();
             return _map.Map<CompanyDetailItemDto>(entity);
         }
         else
         {
-            var entity = await _repo.GetSingleAsync(c => c.Id == id && c.IsDeleted == false, "CompanyIndustries", "CompanyIndustries.Industry");
+            var entity = await _repo.GetSingleAsync(c => c.Id == id && c.IsDeleted == false, "CompanyIndustries", "CompanyIndustries.Industry", "AppUser");
             if (entity == null) throw new NotFoundException<Company>();
             return _map.Map<CompanyDetailItemDto>(entity);
         }
@@ -118,17 +143,23 @@ public class CompanyService : ICompanyService
 
     public async Task UpdateAsync(int id, CompanyUpdateDto dto)
     {
+        if (string.IsNullOrEmpty(_userId)) throw new ArgumentNullException();
+        var user = await _user.Users.Include(a => a.Company).SingleOrDefaultAsync(a => a.Id == _userId);
+        if (user == null) throw new NotFoundException<AppUser>();
+
         if (id < 0) throw new IdIsNegativeException<Company>();
-        var entity = await _repo.FindByIdAsync(id, "CompanyIndustries", "CompanyIndustries.Industry");
+        var entity = await _repo.FindByIdAsync(id, "CompanyIndustries", "CompanyIndustries.Industry", "AppUser");
         if (entity == null) throw new NotFoundException<Company>();
 
         if (await _repo.IsExistAsync(c => c.Name == dto.Name && c.Id != id)) throw new IsExistException<Company>();
+
+        if (entity.AppUserId != _userId) throw new UserNotSameException();
 
         var map = _map.Map(dto, entity);
 
         if (dto.LogoFIle != null)
         {
-           if(entity.Logo != null) _file.Delete(entity.Logo);
+            if (entity.Logo != null) _file.Delete(entity.Logo);
             if (!dto.LogoFIle.IsTypeValid("image")) throw new TypeNotValidException();
             if (!dto.LogoFIle.IsSizeValid(2)) throw new SizeNotValidException();
             entity.Logo = await _file.UploadAsync(dto.LogoFIle, RootContsant.CompanyLogoRoot);
